matching_system/requirements.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Async Task Processing
celery==5.3.6
redis==5.0.1

# Database
asyncpg==0.29.0
pinecone-client==3.0.3

# AI/ML
openai==1.10.0

# WebSocket
python-socketio==5.11.0
aioredis==2.0.1

# Utilities
python-dotenv==1.0.0
